Rename root reducer file to more obvious name
