+++ src/slices/game/game-slice.ts
@@ -247,7 +247,7 @@ export const gameSlice = createSlice({
 
 // Extend the root state typings with this slice.
 //  - The module "name" is actually the relative path to interface definition.
-declare module '..' {
+declare module '../reducer' {
   export interface LazyLoadedSlices extends WithSlice<typeof gameSlice> {}
 }
